Fixed Textures and changed verteces to float

+++ source/load.h
@@ -1,3 +1,3 @@
 
 
-void CreateBlockSide(s16 x1, s16 y1, s16 z1, s16 x2, s16 y2, s16 z2, s16 x3, s16 y3, s16 z3, double u, double v, NE_Material *Material, int Zone, int id);
+void CreateBlockSide(float x1, float y1, float z1, float x2, float y2, float z2, float x3, float y3, float z3, double u, double v, NE_Material *Material, int Zone, int id);
